servicenowintegration/requirements.txt
requests
python-dotenv

pinecone
langchain
langchain-openai
langchain-community
langchain-pinecone

streamlit
beautifulsoup4